requirements.txt
Flask==2.1.2
firebase-admin==5.0.3
requests==2.28.1
gunicorn==20.1.0  # For deployment in production
python-dotenv==0.19.2  # For loading environment variables
PyJWT==2.4.0  # For JWT token management
pydantic==1.9.1  # Optional, for data validation and modeling
pytest==7.1.2  # For running tests
